backend/requirements.txt
# DrawIO AI Backend Dependencies

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
openai>=1.0.0
zhipuai>=2.0.0
redis>=5.0.0
python-multipart>=0.0.6

# MCP 协议通信
mcp>=1.0.0
